+++ app/home/route.ts
@@ -20,6 +20,6 @@ export default class Home extends Route {
 
     @action
     didTransition(this: Home) {
-        this.get('analytics').trackPage();
+        this.analytics.trackPage();
     }
 }